% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cstep.R
\name{cstep}
\alias{cstep}
\title{C-Step of EM algorithm}
\usage{
cstep(postpr)
}
\arguments{
\item{postpr}{(`matrix()`) \cr
The matrix of the posterior probability}
}
\description{
Given the posterior probability, generate a matrix to assign
each individual to a class. The assignment is based on which probability is the largest.
}
