\name{Em-package}
\alias{Em-package}
\alias{Em}
\docType{package}
\title{
\packageTitle{Em}
}
\description{
\packageDescription{Em}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Em}
\packageIndices{Em}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Em}

Maintainer: \packageMaintainer{Em}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
