% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.clogit.R
\name{em.clogit}
\alias{em.clogit}
\title{The em function for `survival::clogit`.}
\usage{
\method{em}{clogit}(
  object,
  latent = 2,
  verbose = FALSE,
  init.method = c("random", "kmeans", "hc"),
  init.prob = NULL,
  algo = c("em", "cem", "sem"),
  cluster.by = NULL,
  max_iter = 500,
  abs_tol = 1e-04,
  concomitant = list(...),
  use.optim = FALSE,
  optim.start = c("random", "sample5"),
  ...
)
}
\arguments{
\item{object}{the model used, e.g. `lm`, `glm`, `gnm`.}

\item{latent}{the number of latent classes.}

\item{verbose}{`True` to print the process of convergence.}

\item{init.method}{the initialization method used in the model.
The default method is `random`. `kmeans` is K-means clustering.
`hc` is model-based agglomerative hierarchical clustering.}

\item{init.prob}{the starting prior probabilities used in classification based method.}

\item{algo}{the algorithm used in em: `em` the default EM algorithm,
the classification em `cem`, or the stochastic em `sem`.}

\item{cluster.by}{a variable to define the level of clustering.}

\item{max_iter}{the maximum iteration for em algorithm.}

\item{abs_tol}{absolute accuracy requested.}

\item{concomitant}{the formula to define the concomitant part of the model.
The default is NULL.}

\item{use.optim}{maximize the complete log likelihood (MLE) by using `optim` and `rcpp` code.The default value is `FALSE`.}

\item{optim.start}{the initialization method of generating the starting value for MLE.}

\item{...}{arguments used in the `model`.}
}
\value{
the fitting object for the model with the class `em`.
}
\description{
The em function for `survival::clogit`.
}
