% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.glmerMod.R
\name{em.glmerMod}
\alias{em.glmerMod}
\title{The em function for glmerMod}
\usage{
\method{em}{glmerMod}(
  object,
  latent = 2,
  verbose = FALSE,
  init.method = c("random", "kmeans", "hc"),
  algo = c("em", "cem", "sem"),
  max_iter = 500,
  concomitant = list(...),
  ...
)
}
\arguments{
\item{object}{the model used, e.g. `lm`, `glm`, `gnm`.}

\item{latent}{the number of latent classes.}

\item{verbose}{`True` to print the process of convergence.}

\item{init.method}{the initialization method used in the model.
The default method is `random`. `kmeans` is K-means clustering.
`hc` is model-based agglomerative hierarchical clustering.}

\item{algo}{the algorithm used in em: `em` the default EM algorithm,
the classification em `cem`, or the stochastic em `sem`.}

\item{max_iter}{the maximum iteration for em algorithm.}

\item{concomitant}{the formula to define the concomitant part of the model.
The default is NULL.}

\item{...}{arguments used in the `model`.}
}
\value{
the fitting object for the model with the class `em`.
}
\description{
The em function for glmerMod
}
