% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM2N.R
\name{EM2N}
\alias{EM2N}
\title{Using EM algorithm}
\usage{
EM2N(p, m_1, m_2)
}
\arguments{
\item{p}{Probability}

\item{m_1}{Expected value of first}

\item{m_2}{Expected value of second}
}
\value{
Iterations towards convergence
}
\description{
Using EM algorithm
}
\examples{
EM2N(0.7, -1, 3)

}
