% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estep.R
\name{estep}
\alias{estep}
\title{E-Step of EM Algorithm}
\usage{
estep(models, pi_matrix)
}
\value{
the fitting result for the model.
}
\description{
This function performs an E-Step of EM Algorithm.
}
\author{
Dongjie Wu
}
