% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.den.R
\name{fit.den.coxph}
\alias{fit.den.coxph}
\title{Fit the density for the survival::clogit}
\usage{
\method{fit.den}{coxph}(object, ...)
}
\arguments{
\item{object}{the fitted model.}

\item{...}{other used arguments.}
}
\value{
the density function.
}
\description{
Fit the density for the survival::clogit
}
