% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.den.R
\name{fit.den.fitdist}
\alias{fit.den.fitdist}
\title{Fitting the density function using in `fitdistrplus::fitdist()`}
\usage{
\method{fit.den}{fitdist}(object, ...)
}
\arguments{
\item{object}{the fitted model.}

\item{...}{other used arguments.}
}
\description{
Fitting the density function using in `fitdistrplus::fitdist()`
}
