% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.den.R
\name{fit.den.glm}
\alias{fit.den.glm}
\title{Fit the density function for a generalized linear regression model.}
\usage{
\method{fit.den}{glm}(object, ...)
}
\arguments{
\item{object}{the fitted model.}

\item{...}{other used arguments.}
}
\description{
Fit the density function for a generalized linear regression model.
}
