% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.den.R
\name{fit.den.glmerMod}
\alias{fit.den.glmerMod}
\title{Fit the density function for a generalized linear mixed effect model.}
\usage{
\method{fit.den}{glmerMod}(object, ...)
}
\arguments{
\item{object}{the fitted model.}

\item{...}{other used arguments.}
}
\value{
the density function.
}
\description{
Fit the density function for a generalized linear mixed effect model.
}
