% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.den.R
\name{fit.den.gnm}
\alias{fit.den.gnm}
\title{Fit the density function for a generalized non-linear regression model.}
\usage{
\method{fit.den}{gnm}(object, ...)
}
\arguments{
\item{object}{the fitted model.}

\item{...}{other used arguments.}
}
\value{
the density function.
}
\description{
Fit the density function for a generalized non-linear regression model.
}
