% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.den.R
\name{fit.den.nnet}
\alias{fit.den.nnet}
\title{Fit the density function for a `nnet` model.}
\usage{
\method{fit.den}{nnet}(object, ...)
}
\arguments{
\item{object}{the fitted model.}

\item{...}{other used arguments.}
}
\value{
the density function.
}
\description{
Fit the density function for a `nnet` model.
}
