% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.den.R
\name{fit.den.plm}
\alias{fit.den.plm}
\title{Fit the density function for a panel regression model.}
\usage{
\method{fit.den}{plm}(object, ...)
}
\arguments{
\item{object}{the fitted model.}

\item{...}{other used arguments.}
}
\description{
Fit the density function for a panel regression model.
}
