% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.den.R
\name{fit.den}
\alias{fit.den}
\title{Fit the density function for a fitted model.}
\usage{
fit.den(object, ...)
}
\arguments{
\item{object}{the fitted model such as `lm`.}

\item{...}{other used arguments.}
}
\description{
This function generates the probability density of given models.
}
