% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten}
\alias{flatten}
\title{Flatten a data.frame or matrix by column or row with its name.
The name will be transformed into the number of row/column plus
the name of column/row separated by `.`.}
\usage{
flatten(x, by = c("col", "row"))
}
\arguments{
\item{x}{a data.frame or matrix.}

\item{by}{either by column or by row.}
}
\value{
a flattened vector with names
}
\description{
Flatten a data.frame or matrix by column or row with its name.
The name will be transformed into the number of row/column plus
the name of column/row separated by `.`.
}
