% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.em.R
\name{init.em.hc}
\alias{init.em.hc}
\title{model-based agglomerative hierarchical clustering}
\usage{
\method{init.em}{hc}(object, ...)
}
\arguments{
\item{object}{A matrix.}

\item{...}{other used arguments.}
}
\description{
model-based agglomerative hierarchical clustering
}
