% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.em.R
\name{init.em.kmeans}
\alias{init.em.kmeans}
\title{K-mean initialization}
\usage{
\method{init.em}{kmeans}(object, ...)
}
\arguments{
\item{object}{A matrix.}

\item{...}{other used arguments.}
}
\description{
K-mean initialization
}
