% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.em.R
\name{init.em.random}
\alias{init.em.random}
\title{Random initialization}
\usage{
\method{init.em}{random}(object, ...)
}
\arguments{
\item{object}{A matrix.}

\item{...}{other used arguments.}
}
\value{
The posterior probability matrix
}
\description{
Random initialization
}
