% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.em.R
\name{init.em.random.weights}
\alias{init.em.random.weights}
\title{Random initialization with weights}
\usage{
\method{init.em}{random.weights}(object, ...)
}
\arguments{
\item{object}{A matrix.}

\item{...}{other used arguments.}
}
\value{
The posterior probability matrix
}
\description{
Random initialization with weights
}
