% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.em.R
\name{init.em}
\alias{init.em}
\title{Initialization of EM algorithm}
\usage{
init.em(object, ...)
}
\arguments{
\item{object}{A matrix.}

\item{...}{other used arguments.}
}
\value{
The posterior probability matrix
}
\description{
Given a matrix with number of rows equal to the number of
observation and number of columns equal to the number of latent classes,
function `init.em` generate the posterior probability using that matrix
based on the method set by the user.
}
