% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.em.R
\name{init.em.sample5}
\alias{init.em.sample5}
\title{Initialization using sampling 5 times.}
\usage{
\method{init.em}{sample5}(object, ...)
}
\arguments{
\item{object}{A matrix.}

\item{...}{other used arguments.}
}
\value{
The posterior probability matrix
}
\description{
Initialization using sampling 5 times.
}
