% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.coxph.R
\name{model.matrix.coxph}
\alias{model.matrix.coxph}
\title{This function is from `survival` package.}
\usage{
\method{model.matrix}{coxph}(object, data = NULL, contrast.arg = object$contrasts, ...)
}
\arguments{
\item{object}{a model frame.}

\item{data}{new data set.}

\item{contrast.arg}{the argments of contrasts.}

\item{...}{other arguments.}
}
\description{
This function is from `survival` package.
}
