% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.concomitant.R
\name{mstep.concomitant}
\alias{mstep.concomitant}
\title{The mstep for the concomitant model.}
\usage{
mstep.concomitant(formula, data, postpr)
}
\arguments{
\item{formula}{the formula of the concomitant model.}

\item{data}{the data or model.frame related to the concomitant model.}

\item{postpr}{the posterior probability matrix.}
}
\value{
the function returns a fitted nnet object.
}
\description{
This section was inspired by Flexmix.
}
