% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.concomitant.R
\name{mstep.concomitant.refit}
\alias{mstep.concomitant.refit}
\title{The refit of for the concomitant model.
This section was inspired by Flexmix.}
\usage{
mstep.concomitant.refit(formula, data, postpr)
}
\arguments{
\item{formula}{the formula of the concomitant model.}

\item{data}{the data or model.frame related to the concomitant model.}

\item{postpr}{the posterior probability matrix.}
}
\value{
the function returns a fitted multinom object.
}
\description{
The refit of for the concomitant model.
This section was inspired by Flexmix.
}
