% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.em.R
\name{multi.em.default}
\alias{multi.em.default}
\title{Default generic for multi.em}
\usage{
\method{multi.em}{default}(object, iter = 10, parallel = FALSE, ...)
}
\arguments{
\item{object}{the model to use in em, e.g. `lm`, `glm`, `gnm`}

\item{iter}{number of iterations for running EM algorithm.}

\item{parallel}{whether to use the parallel computing.}

\item{...}{arguments used in em.}
}
\value{
return the `em` object with the maximum log-likelihood.
}
\description{
Default generic for multi.em
}
