% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.em.R
\name{multi.em}
\alias{multi.em}
\title{Multiple run of EM algorithm}
\usage{
multi.em(object, ..., iter = 10, parallel = FALSE)
}
\arguments{
\item{object}{the model to use in em, e.g. `lm`, `glm`, `gnm`}

\item{...}{arguments in the model object.}

\item{iter}{number of iterations for running EM algorithm.}

\item{parallel}{whether to use the parallel computing.}
}
\value{
return the `em` object with the maximum log-likelihood.
}
\description{
Multiple run of EM algorithm
}
