% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2density.R
\name{plot2density}
\alias{plot2density}
\title{Plot density}
\usage{
plot2density(df)
}
\arguments{
\item{df}{Two factor data frame each 200 from rnorm(m_1,1).}
}
\value{
Density plot over histogram
}
\description{
Plot density
}
\examples{
 set.seed(1)
library(ggplot2)
m_1<-1
m_2<-3
df <- data.frame(
sex=factor(rep(c("F", "M"), each=200)),
 weight=round(c(rnorm(200, mean=m_1, sd=1),
                rnorm(200, mean=m_2, sd=1)))
)
plot2density(df)
}
