% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{print.em}
\alias{print.em}
\title{Print the `em` object}
\usage{
\method{print}{em}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{the `em` object.}

\item{digits}{the maximum digits printed, the default is `3L`.}

\item{...}{other arguments used.}
}
\value{
print the `em` object on the screen.
}
\description{
Print the `em` object
}
