% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{print.summary.em}
\alias{print.summary.em}
\title{Print the `summary.em` object}
\usage{
\method{print}{summary.em}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{the `summary.em` object.}

\item{digits}{the maximum digits printed, the default is `3L`.}

\item{signif.stars}{logical; if `TRUE`, P-values are additionally encoded visually 
as `significance stars` in order to help scanning of long coefficient tables. 
It defaults to the `show.signif.stars` slot of options.}

\item{...}{other augments used in `printCoefmat`.}
}
\value{
print the `summary.em` object on the screen.
}
\description{
Print the `summary.em` object
}
