% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simbinom}
\alias{simbinom}
\title{Simulated Data from a logistic regression}
\format{
A data frame with 10000 rows and 2 variables:
\describe{
  \item{y}{A dependent variable generaged from a mixture of a logistic regression with x}
  \item{x}{An independent variable}
}
}
\source{
<https://www.github.com/wudongjie/em>
}
\usage{
simbinom
}
\description{
A data set with simulated data from a mixture of a logistic regression.
}
\keyword{data}
