% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simclogit}
\alias{simclogit}
\title{Simulated Data from a conditional logistic regression}
\format{
A data frame with 10000 rows and 4 variables:
\describe{
  \item{y1}{A dependent variable generaged from a conditional logistic regression with x2 and x3}
  \item{y2}{A dependent variable generaged from a mixture of two classes of conditional logistic regressions with x2 and x3}
  \item{x2}{A dummy variable showing whether x is equal to level 2}
  \item{x3}{A dummy variable showing whether x is equal to level 3}
}
}
\source{
<https://www.github.com/wudongjie/em>
}
\usage{
simclogit
}
\description{
A data set with simulated data from a mixture of a conditional logistic regression.
}
\keyword{data}
