% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simreg}
\alias{simreg}
\title{Simulated Regression Data}
\format{
A data frame with 1000 rows and 5 variables:
\describe{
  \item{yp}{A dependent variable generaged from a mixture of a poisson regression with x}
  \item{yn}{A dependent variable generaged from a mixture of a linear regression with x}
  \item{yc}{A dependent variable generaged from a mixture of a linear regression with x and a concomitant variable of z}
  \item{x}{An independent variable}
  \item{z}{A concomitant variable}
}
}
\source{
<https://www.github.com/wudongjie/em>
}
\usage{
simreg
}
\description{
A data set with simulated data from mixture regression models.
}
\keyword{data}
