% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstep.R
\name{sstep}
\alias{sstep}
\title{S-step of EM algorithm}
\usage{
sstep(postpr)
}
\arguments{
\item{postpr}{(`matrix()`) \cr
The matrix of the posterior probability}
}
\description{
Given the posterior probability, generate a matrix to assign
each individual to a class. The assignment is randomly sampled based on the posterior probability.
}
