% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vdummy}
\alias{vdummy}
\title{Transform a factor variable to a matrix of dummy variables}
\usage{
vdummy(x)
}
\arguments{
\item{x}{a factor vector}
}
\value{
a matrix of dummy variables
}
\description{
Transform a factor variable to a matrix of dummy variables
}
