\name{EM-package}
\alias{EM-package}
\alias{EM}
\docType{package}
\title{
\packageTitle{EM}
}
\description{
\packageDescription{EM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EM}
\packageIndices{EM}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{EM}

Maintainer: \packageMaintainer{EM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
