% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{em.default}
\alias{em.default}
\title{The default em function}
\usage{
\method{em}{default}(
  object,
  latent = 2,
  verbose = F,
  init.method = c("random", "kmeans", "hc"),
  init.prob = NULL,
  algo = c("em", "cem", "sem"),
  cluster.by = NULL,
  max_iter = 500,
  abs_tol = 1e-04,
  concomitant = list(...),
  use.optim = F,
  optim.start = c("random", "sample5"),
  ...
)
}
\arguments{
\item{object}{the model used, e.g. `lm`, `glm`, `gnm`.}

\item{latent}{the number of latent classes.}

\item{verbose}{`True` to print the process of convergence.}

\item{init.method}{the initialization method used in the model.
The default method is `random`. `kmeans` is K-means clustering. 
`hc` is model-based agglomerative hierarchical clustering.}

\item{init.prob}{the starting prior probabilities used in classification based method.}

\item{algo}{the algorithm used in em: `em` the default EM algorithm, 
the classification em `cem`, or the stochastic em `sem`.}

\item{max_iter}{the maximum iteration for em algorithm.}

\item{concomitant}{the formula to define the concomitant part of the model.
The default is NULL.}

\item{...}{arguments used in the `model`.}
}
\value{
the fitting object for the model with the class `em`.
}
\description{
The default em function
}
