% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{em}
\alias{em}
\title{A Generic EM Algorithm}
\usage{
em(object, ...)
}
\description{
A generic EM algorithm that can work on specific models/classes,
models/classes that can use `logLik` to extract the log-likelihood function.
}
\author{
Dongjie Wu
}
