% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{em}
\alias{em}
\title{A Generic EM Algorithm}
\usage{
em(
  model,
  ...,
  latent = 2,
  verbose = F,
  init.method = c("random", "kmeans"),
  max_iter = 500,
  concomitant = NULL
)
}
\value{
the fitting result for the model.
}
\description{
A generic EM algorithm that can work on specific models/classes,
models/classes that can use \code{logLik} to extract the log-likelihood function.
}
\author{
Dongjie Wu
}
