% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{em}
\alias{em}
\title{A Generic EM Algorithm}
\usage{
em(object, ...)
}
\arguments{
\item{object}{the model used, e.g. `lm`, `glm`, `gnm`, `clogit`, `multinom`}

\item{...}{arguments used in the `model`.}
}
\description{
This is a generic EM algorithm that can work on specific models/objects.
Currently, it supports `lm`, `glm`, `gnm` in package gnm,
`clogit` in package survival and `multinom` in package nnet.
Use `?em.default` to check the manual of the default function of `em`.
}
\author{
Dongjie Wu
}
