% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{em}
\alias{em}
\title{A Generic EM Algorithm}
\usage{
em(object, ...)
}
\arguments{
\item{object}{the model used, e.g. `lm`, `glm`, `gnm`.}

\item{...}{arguments used in the `model`.}
}
\description{
A generic EM algorithm that can work on specific models/objects.
}
\author{
Dongjie Wu
}
