% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.em.R
\name{logLik.em}
\alias{logLik.em}
\title{This function computes logLik of EM Algorithm.}
\usage{
\method{logLik}{em}(object, ...)
}
\arguments{
\item{object}{an object of `em`.}

\item{...}{other used arguments.}
}
\value{
the log-likelihood value
}
\description{
This function computes logLik of EM Algorithm.
}
