% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R
\name{mstep}
\alias{mstep}
\title{M-Step of EM Algorithm}
\usage{
mstep(models, pars, post_pr = NULL)
}
\value{
the fitting result for the model.
}
\description{
This function performs an M-Step of EM Algorithm.
}
\author{
Dongjie Wu
}
