% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.em.R
\name{multi.em}
\alias{multi.em}
\title{Multiple run of EM algorithm}
\usage{
multi.em(object, ...)
}
\arguments{
\item{object}{the model to use in em, e.g. `lm`, `glm`, `gnm`}

\item{...}{arguments used in em.}
}
\value{
return the `em` object with the maximum log-likelihood.
}
\description{
Multiple run of EM algorithm
}
