% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{A Generic SEM Algorithm}
\usage{
sem(model, pars, latent = 2, verbose = F, max_iter = 500)
}
\value{
the fitting result for the model.
}
\description{
A generic SEM algorithm that can work on specific models/classes,
models/classes that can use \code{logLik} to extract the log-likelihood function.
}
\author{
Dongjie Wu
}
