% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simclogit}
\alias{simclogit}
\title{Simulated Data from a conditional logistic regression}
\format{
A data frame with 10000 rows and 4 variables:
\describe{
  \item{x2}{A dummy variable showing whether x is equal to level 2}
  \item{x3}{A dummy variable showing whether x is equal to level 3}
  \item{a2}{Whether the alternative choice 2 is chosen}
  \item{a2_x2}{Interaction between a2 and x2}
  \item{a2_x3}{Interaction between a2 and x3}
  \item{a3}{Whether the alternative choice 3 is chosen}
  \item{a3_x2}{Interaction between a3 and x2}
  \item{a3_x3}{Interaction between a3 and x3}
  \item{chosen1}{Whether the observation-alternative combination is chosen (Generated by a one-class regression).}
  \item{chosen2}{Whether the observation-alternative combination is chosen (Generated by a two-class mixed regression).}
  \item{fid}{Family ID}
  \item{id}{Individual ID}
  \item{z}{Other variables}
}
}
\source{
<https://www.github.com/wudongjie/em>
}
\usage{
simclogit
}
\description{
A data set with simulated data from a mixture of a conditional logistic regression.
}
\keyword{data}
